Switch param input to JSON stdin only
By only allowing taking stdin / json data, we can now add the 'additional_params' escape hatch

Any not specified Nutanix V3 api data not created in the task itself can now be passed as additional_params
and it will be merged with the internal data structure in the task, and allowing it to be overwritten by
data in the task as specific params. This allows for a 'generic' set of stuff to be passed via this
field in an api call, and then specify the more specific stuff (ie, machine name) in that specific param field.